Create pre-build.xml to setup Env
At moment, before the pkg-list installation, there is no chance that
some operetions can be done to the filesystem. Thus, the cmdline.txt
and config.txt can not be created before raspberrypi-sys-mods gets
installed(which depends on the files).

A workaround for this can be, create the files in the finetuning of
a building, and they are inherented by the following building and
generate the target image.