Add the update-mime-database to revpi-factory-reset
Due to unknown reason(s), the update of mime failed for raspberrypi-ui-mods
during installation in ELBE env. But this can be done successfully when the
image running on a device.

As a workaround, adding the command to revpi-factory-reset.

The turning of the pkg-list might be not needed, but in this order it works.

Signed-off-by: Zhi Han <z.han@kunbus.com>